third_party/WebKit/LayoutTests/fast/dom/Range/range-insertNode-splittext-expected.txt

Test for proper behavior of Range.insertNode when splitting text nodes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS p.childNodes.length is 3
PASS p.childNodes[0] is t1
PASS p.childNodes[0].data is "12"
PASS p.childNodes[1] is t2
PASS p.childNodes[1].data is "ABCDE"
PASS p.childNodes[2].data is "345"
PASS r.collapsed is false
PASS r.commonAncestorContainer is p
PASS r.startContainer is t1
PASS r.startOffset is 2
PASS r.endContainer is t3
PASS r.endOffset is 1
PASS r.toString() is "ABCDE3"
PASS successfullyParsed is true

TEST COMPLETE
